+++ lib/generators/rodauth/templates/app/mailers/rodauth_mailer.rb
@@ -1,64 +1,60 @@
 class RodauthMailer < ApplicationMailer
-  def verify_account(name = nil, account_id, key)
-    @email_link = email_link(name, :verify_account, account_id, key)
-    @account = find_account(name, account_id)
+  def verify_account(name, account_id, key)
+    @rodauth = rodauth(name, account_id) { @verify_account_key_value = key }
+    @account = @rodauth.rails_account
 
-    mail to: @account.email, subject: rodauth(name).verify_account_email_subject
+    mail to: @account.email, subject: @rodauth.verify_account_email_subject
   end
 
-  def reset_password(name = nil, account_id, key)
-    @email_link = email_link(name, :reset_password, account_id, key)
-    @account = find_account(name, account_id)
+  def reset_password(name, account_id, key)
+    @rodauth = rodauth(name, account_id) { @reset_password_key_value = key }
+    @account = @rodauth.rails_account
 
-    mail to: @account.email, subject: rodauth(name).reset_password_email_subject
+    mail to: @account.email, subject: @rodauth.reset_password_email_subject
   end
 
-  def verify_login_change(name = nil, account_id, key)
-    @email_link = email_link(name, :verify_login_change, account_id, key)
-    @account = find_account(name, account_id)
+  def verify_login_change(name, account_id, key)
+    @rodauth = rodauth(name, account_id) { @verify_login_change_key_value = key }
+    @account = @rodauth.rails_account
     @new_email = @account.login_change_key.login
 
-    mail to: @new_email, subject: rodauth(name).verify_login_change_email_subject
+    mail to: @new_email, subject: @rodauth.verify_login_change_email_subject
   end
 
-  def password_changed(name = nil, account_id)
-    @account = find_account(name, account_id)
+  def password_changed(name, account_id)
+    @rodauth = rodauth(name, account_id)
+    @account = @rodauth.rails_account
 
-    mail to: @account.email, subject: rodauth(name).password_changed_email_subject
+    mail to: @account.email, subject: @rodauth.password_changed_email_subject
   end
 
-  # def email_auth(name = nil, account_id, key)
-  #   @email_link = email_link(name, :email_auth, account_id, key)
-  #   @account = find_account(name, account_id)
+  # def reset_password_notify(name, account_id)
+  #   @rodauth = rodauth(name, account_id)
+  #   @account = @rodauth.rails_account
 
-  #   mail to: @account.email, subject: rodauth(name).email_auth_email_subject
+  #   mail to: @account.email, subject: @rodauth.reset_password_notify_email_subject
   # end
 
-  # def unlock_account(name = nil, account_id, key)
-  #   @email_link = email_link(name, :unlock_account, account_id, key)
-  #   @account = find_account(name, account_id)
+  # def email_auth(name, account_id, key)
+  #   @rodauth = rodauth(name, account_id) { @email_auth_key_value = key }
+  #   @account = @rodauth.rails_account
 
-  #   mail to: @account.email, subject: rodauth(name).unlock_account_email_subject
+  #   mail to: @account.email, subject: @rodauth.email_auth_email_subject
+  # end
+
+  # def unlock_account(name, account_id, key)
+  #   @rodauth = rodauth(name, account_id) { @unlock_account_key_value = key }
+  #   @account = @rodauth.rails_account
+
+  #   mail to: @account.email, subject: @rodauth.unlock_account_email_subject
   # end
 
   private
 
-  def find_account(_name, account_id)
-<% if defined?(ActiveRecord::Railtie) -%>
-    Account.find(account_id)
-<% else -%>
-    Account.with_pk!(account_id)
-<% end -%>
-  end
-
-  def email_link(name, action, account_id, key)
-    instance = rodauth(name)
-    instance.instance_variable_set(:@account, { id: account_id })
-    instance.instance_variable_set(:"@#{action}_key_value", key)
-    instance.public_send(:"#{action}_email_link")
-  end
-
-  def rodauth(name)
-    RodauthApp.rodauth(name).allocate
+  def rodauth(name, account_id, &block)
+    instance = RodauthApp.rodauth(name).allocate
+    instance.instance_eval { @account = account_ds(account_id).first! }
+    instance.instance_eval(&block) if block
+    instance
   end
 end