+++ lib/rodauth/rails/feature/instrumentation.rb
@@ -2,6 +2,8 @@ module Rodauth
   module Rails
     module Feature
       module Instrumentation
+        extend ActiveSupport::Concern
+
         private
 
         def _around_rodauth
@@ -32,7 +34,7 @@ module Rodauth
           request = rails_request
 
           raw_payload = {
-            controller: scope.class.superclass.name,
+            controller: self.class.roda_class.name,
             action: "call",
             request: request,
             params: request.filtered_parameters,