+++ test/integration/controller_test.rb
@@ -10,12 +10,6 @@ class ControllerTest < IntegrationTest
   test "current account" do
     register(login: "user@example.com", verify: true)
     assert_text "Authenticated as user@example.com"
-
-    capture_io { Account.last.destroy }
-
-    visit "/"
-    assert_text "Please login to continue"
-    assert_equal "/login", page.current_path
   end
 
   test "executing controller methods" do
@@ -33,5 +27,5 @@ class ControllerTest < IntegrationTest
     assert_equal 302, page.status_code
     assert_equal "/", page.response_headers["Location"]
     assert_equal "true", page.response_headers["X-After"]
-  end unless ActiveRecord::VERSION::MAJOR <= 4 && RUBY_ENGINE == "jruby"
+  end
 end