+++ test/rails_app/app/misc/rodauth_app.rb
@@ -28,9 +28,10 @@ class RodauthApp < Rodauth::Rails::App
     r.on("json") { r.rodauth(:json) }
 
     r.on("assets") { "" }
+    r.get("admin/custom") { "Custom admin route" }
 
     if r.path == rails_routes.auth1_path
-      rodauth.require_authentication
+      rodauth.require_account
     end
   end
 end