+++ test/rails_app/app/misc/rodauth_main.rb
@@ -32,5 +32,8 @@ class RodauthMain < Rodauth::Rails::Auth
     logout_redirect { rails_routes.root_path }
     verify_account_redirect { login_redirect }
     reset_password_redirect { login_path }
+    title_instance_variable :@page_title
+
+    verify_login_change_route nil
   end
 end